% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{class_constructor}
\alias{class_constructor}
\title{class constructor function}
\usage{
class_constructor(x)
}
\arguments{
\item{x}{object for class creation}
}
\description{
class constructor function
}
