% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{linreg method for coef()}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{input of the class \code{linreg} containing model data}
}
\description{
linreg method for coef()
}
