% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_linreg.R
\name{coef}
\alias{coef}
\title{Generic Method for coef()}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{input containing model data}

\item{...}{additional input to the coef function}
}
\description{
Generic Method for coef()
}
