% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{linreg Method for plot()}
\usage{
\method{plot}{linreg}(x, y, ...)
}
\arguments{
\item{x}{object containing model data to be used for plotting}

\item{y}{the y coordinates for plotting}

\item{...}{other parameters for plotting}
}
\description{
linreg Method for plot()
}
