% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot}
\alias{plot}
\title{Generic Method for plot()}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{object containing model data to be used for plotting}

\item{y}{the y coordinates for plotting}
}
\description{
Generic Method for plot()
}
