% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{linreg method for pred()}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{input of the class \code{linreg} containing model data}

\item{...}{additional input to pred function}
}
\description{
linreg method for pred()
}
