% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_linreg.R
\name{pred}
\alias{pred}
\title{Generic Method for pred()}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{input containing model data}

\item{...}{additional input to pred function}
}
\description{
Generic Method for pred()
}
