% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{linreg method for print()}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{input of the class \code{linreg}}
}
\description{
linreg method for print()
}
