% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_linreg.R
\name{print}
\alias{print}
\title{Generic Method for print()}
\usage{
print(x, ...)
}
\arguments{
\item{x}{input to be printed of the class linreg}

\item{...}{other inputs to be printed}
}
\description{
Generic Method for print()
}
