% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{linreg method for resid()}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{input of the class \code{linreg} containing model data}
}
\description{
linreg method for resid()
}
