% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_linreg.R
\name{resid}
\alias{resid}
\title{Generic Method for resid()}
\usage{
resid(object, ...)
}
\arguments{
\item{object}{input containing model data}

\item{...}{other inputs to the resid function}
}
\description{
Generic Method for resid()
}
