% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linreg.R
\name{summary}
\alias{summary}
\title{Generic Method for summary()}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{input containing model data}

\item{...}{other inputs to summary function}
}
\description{
Generic Method for summary()
}
