% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{linreg method for summary()}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{input of the class \code{linreg} containing model data}

\item{...}{other inputs to summary function}
}
\description{
linreg method for summary()
}
